refactor(capture): move lexical sync helpers to capture-sync.ts
Pure move of syncCapturedFile, ensureCapturedFileIndexed,
CaptureSyncError and their types out of the 1100-line capture.ts;
capture.ts re-exports them so every import path keeps working.
